Move sanitize_value to be common, Fix InPassportRecognizer  (#1519)
* Move sanitize_value to be common
Fix Indian Passport Number regex

* fix regex

* export module

* export module

* fix imports

* move to entity_recognizer.py

* move to entity_recognizer.py
