+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/in_aadhaar_recognizer.py
@@ -1,7 +1,6 @@
 from typing import List, Optional, Tuple
 
-from presidio_analyzer import Pattern, PatternRecognizer
-from presidio_analyzer.analyzer_utils import PresidioAnalyzerUtils as Utils
+from presidio_analyzer import EntityRecognizer, Pattern, PatternRecognizer
 
 
 class InAadhaarRecognizer(PatternRecognizer):
@@ -58,7 +57,9 @@ class InAadhaarRecognizer(PatternRecognizer):
 
     def validate_result(self, pattern_text: str) -> bool:
         """Determine absolute value based on calculation."""
-        sanitized_value = Utils.sanitize_value(pattern_text, self.replacement_pairs)
+        sanitized_value = EntityRecognizer.sanitize_value(
+            pattern_text, self.replacement_pairs
+        )
         return self.__check_aadhaar(sanitized_value)
 
     def __check_aadhaar(self, sanitized_value: str) -> bool:
@@ -67,8 +68,60 @@ class InAadhaarRecognizer(PatternRecognizer):
             len(sanitized_value) == 12
             and sanitized_value.isnumeric() is True
             and int(sanitized_value[0]) >= 2
-            and Utils.is_verhoeff_number(int(sanitized_value)) is True
-            and Utils.is_palindrome(sanitized_value) is False
+            and self._is_verhoeff_number(int(sanitized_value)) is True
+            and self._is_palindrome(sanitized_value) is False
         ):
             is_valid_aadhaar = True
         return is_valid_aadhaar
+
+    @staticmethod
+    def _is_palindrome(text: str, case_insensitive: bool = False):
+        """
+        Validate if input text is a true palindrome.
+
+        :param text: input text string to check for palindrome
+        :param case_insensitive: optional flag to check palindrome with no case
+        :return: True / False
+        """
+        palindrome_text = text
+        if case_insensitive:
+            palindrome_text = palindrome_text.replace(" ", "").lower()
+        return palindrome_text == palindrome_text[::-1]
+
+    @staticmethod
+    def _is_verhoeff_number(input_number: int):
+        """
+        Check if the input number is a true verhoeff number.
+
+        :param input_number:
+        :return:
+        """
+        __d__ = [
+            [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+            [1, 2, 3, 4, 0, 6, 7, 8, 9, 5],
+            [2, 3, 4, 0, 1, 7, 8, 9, 5, 6],
+            [3, 4, 0, 1, 2, 8, 9, 5, 6, 7],
+            [4, 0, 1, 2, 3, 9, 5, 6, 7, 8],
+            [5, 9, 8, 7, 6, 0, 4, 3, 2, 1],
+            [6, 5, 9, 8, 7, 1, 0, 4, 3, 2],
+            [7, 6, 5, 9, 8, 2, 1, 0, 4, 3],
+            [8, 7, 6, 5, 9, 3, 2, 1, 0, 4],
+            [9, 8, 7, 6, 5, 4, 3, 2, 1, 0],
+        ]
+        __p__ = [
+            [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+            [1, 5, 7, 6, 2, 8, 3, 0, 9, 4],
+            [5, 8, 0, 3, 7, 9, 6, 1, 4, 2],
+            [8, 9, 1, 6, 0, 4, 3, 5, 2, 7],
+            [9, 4, 5, 3, 1, 2, 6, 8, 7, 0],
+            [4, 2, 8, 6, 5, 7, 3, 9, 0, 1],
+            [2, 7, 9, 3, 8, 0, 6, 4, 1, 5],
+            [7, 0, 4, 6, 9, 1, 3, 2, 5, 8],
+        ]
+        __inv__ = [0, 4, 3, 2, 1, 5, 6, 7, 8, 9]
+
+        c = 0
+        inverted_number = list(map(int, reversed(str(input_number))))
+        for i in range(len(inverted_number)):
+            c = __d__[c][__p__[i % 8][inverted_number[i]]]
+        return __inv__[c] == 0
